Abstract Typefully box util
